==> Part_1_Data_Parsing.ipynb <==
# Part 1: Data Parsing

In this part of the tutorials, you will be guided through the different parsing steps needed to ensure high-quality consistent data formats. <br>
<br>
This includes: 
- Loading the data
- Column selection
- Remove entries with missing data
- Remove CDR3β sequences containing non-amino acid characters
- Remove CDR3β sequences not starting/ending with the conserved cysteine (C) and end with a phenylalanine (F) residue
- Merging and excluding duplicate sequences from the data


# Import packages

import warnings
warnings.filterwarnings('ignore')

import pandas as pd

from pyteomics import parser

# Load the data

# Define your current working directory
cwd = '/your/working/directory'

# Read in the correct data files
Patient_1_pre = pd.read_csv(f'{cwd}/Data/HN005_TUMOR_PRE.tsv', sep='\t')
Patient_1_post = pd.read_csv(f'{cwd}/Data/HN005_TUMOR_POST.tsv', sep='\t')

Patient_2_pre = pd.read_csv(f'{cwd}/Data/HN006_TUMOR_PRE.tsv', sep='\t')
Patient_2_post = pd.read_csv(f'{cwd}/Data/HN006_TUMOR_POST.tsv', sep='\t')

Patient_3_pre = pd.read_csv(f'{cwd}/Data/HN009_TUMOR_PRE.tsv', sep='\t')
Patient_3_post = pd.read_csv(f'{cwd}/Data/HN009_TUMOR_POST.tsv', sep='\t')

Patient_3_pre

# Define parsing functions

def get_columns(data):

    # Extract informative columns from the large dataframes
    result = data[['aminoAcid','vFamilyName', 'jFamilyName',  'count (templates/reads)', 'frequencyCount (%)', 'sequenceStatus']]

    # Include only the productive TCRs as described by Immuneaccess
    result = result.loc[result['sequenceStatus'] == 'In']

    # Change column names and add a column describing the full sequence (V- and J-family included)
    result.rename(columns={'aminoAcid': 'junction_aa', 'count (templates/reads)': 'Clone_count', 'frequencyCount (%)': 'Clone_frequency', 'vFamilyName': 'v_call', 'jFamilyName': 'j_call'}, inplace=True)
    result['Full_CDR3'] = result['junction_aa'] + '_' + result['v_call'] + '_' + result['j_call']
    result = result.dropna()

    return result


# Define all known amino acids
def is_amino_acid(cdr3):
    return all(aa in parser.std_amino_acids for aa in cdr3)


def parse_data(dataframe):

    # Select only the sequences with known amino acids
    Step1 = dataframe[dataframe['junction_aa'].apply(is_amino_acid)]

    # Make sure each sequence starts with C and ends with F
    start_c = Step1['junction_aa'].str.startswith('C', na=False)
    end_f = Step1['junction_aa'].str.endswith('F', na=False)
    parse_result = Step1[start_c & end_f].reset_index(drop=True)

    # Sum the clone counts for duplicate TCRs and merge into one entry
    parse_result['Total_count'] = parse_result.groupby(['junction_aa','v_call','j_call'])['Clone_count'].transform('sum')
    parse_result['Total_frequency'] = parse_result.groupby(['junction_aa','v_call','j_call'])['Clone_frequency'].transform('sum')
    parse_result = parse_result[['junction_aa', 'v_call', 'j_call', 'Full_CDR3', 'Total_count', 'Total_frequency']]
    parse_result.drop_duplicates(subset=['junction_aa', 'v_call', 'j_call'], inplace=True)

    return parse_result


# Apply column extraction and data parsing on dataframe
def full_parsing(data, patient):
    Columns = get_columns(data)
    Parsing = parse_data(Columns)
    print(f'Parsing: {patient}')
    print(f'The number of TCRs before parsing is: {len(Columns)}')
    print(f'The number of TCRs after parsing is: {len(Parsing)}')
    print(f'The number of TCRs removed is : {len(Columns) - len(Parsing)} \n')

    return Parsing


# Get total number of TCRs per patient
def get_patient_total(data_pre, data_post, Patient):
    data_pre = pd.DataFrame(data_pre['Full_CDR3'])
    data_post = pd.DataFrame(data_post['Full_CDR3'])
    merged = pd.concat([data_pre,data_post]).drop_duplicates().reset_index(drop=True)
    print('The total number of unique sequences for', Patient, 'is:', len(merged))

    return merged

### Apply the parsing function defined above on the TCR data

# Parse data

# Execute parsing steps for each dataset
Data = [Patient_1_pre, Patient_1_post, Patient_2_pre, Patient_2_post, Patient_3_pre, Patient_3_post]
Data_names = ['Patient_1_pre', 'Patient_1_post', 'Patient_2_pre', 'Patient_2_post', 'Patient_3_pre', 'Patient_3_post']
Parsed_frames = []

for i,j in zip(Data,Data_names):
    res = full_parsing(i,j)
    Parsed_frames.append(res)

P1_pre, P1_post, P2_pre, P2_post, P3_pre, P3_post = [pd.DataFrame(j) for j in Parsed_frames]

# Get total number of unique sequences for each patient
Patient1_total = get_patient_total(P1_pre, P1_post, 'Patient 1')
Patient2_total = get_patient_total(P2_pre, P2_post, 'Patient 2')
Patient3_total = get_patient_total(P3_pre, P3_post, 'Patient 3')

# Example of resulting dataframe

P1_pre

# Export parsed data to file

Dataframes = [P1_pre, P1_post, P2_pre, P2_post, P3_pre, P3_post, Patient1_total, Patient2_total, Patient3_total]
Names=['P1_pre', 'P1_post', 'P2_pre', 'P2_post', 'P3_pre', 'P3_post', 'Patient1_total', 'Patient2_total', 'Patient3_total']

for i,j in zip(Dataframes, Names):
    i.to_csv(f'{cwd}/Processed_data/{j}_data.tsv', sep='\t', index=False)

# Export part of parsed data to files for clustering

Dataframes = [P1_pre, P1_post, P2_pre, P2_post, P3_pre, P3_post]
Names=['P1_pre', 'P1_post', 'P2_pre', 'P2_post', 'P3_pre', 'P3_post']

for i,j in zip(Dataframes, Names):
    i = i[['junction_aa', 'v_call', 'j_call']]
    i.to_csv(f'{cwd}/Processed_data/{j}_cluster_data.tsv', sep='\t', index=False)
